Update README.md
Replace 'growth' with 'grounding and understanding' to align with Live Agile principle

I want to avoid using the term 'growth.' I feel there is too much emphasis on 'growth for growth's sake' in the prevailing culture. Reflecting on natural patterns, cancer serves as a metaphor for parasitical growth that ultimately destroys its host. In the spirit of deeper connection and introspection, I have replaced this term with 'grounding and understanding' to better reflect Live Agile's philosophy.